requirements.txt
aiogram~=3.0.0b7
psycopg2-binary
dj_database_url~=2.0.0
Django~=4.2
django-split-settings~=1.2.0
django-stubs~=4.2.1
gunicorn~=20.1.0
python-dotenv~=1.0.0
starlette~=0.28.0
uvicorn~=0.22.0
django-guardian
redis
djangorestframework


# req_dev.txt
black~=23.3.0
flakeheaven~=3.3.0
isort~=5.12.0
mypy~=1.4
perflint~=0.7.3
pip-check-updates~=0.22.0
pre-commit~=3.3.3
pylint~=2.17.4
pytest~=7.4.0
pytest-cov~=4.1.0
